apisample7.html

<!DOCTYPE HTML>
<title>Example with file_is_test (should fail)</title>
<script src="testharness.js"></script>
<script src="testharnessreport.js"></script>
<script>setup({file_is_test:true})

onload = function() {
  assert_true(false);
  done();
}
</script>
<div id=log></div>
